moved auth screen logic into dialog service
