+++ src/app/core/firebase/components/firebase-auth-dialog/firebase-auth-dialog.component.ts
@@ -1,18 +1,16 @@
-import { Component, computed, effect, inject, signal, ViewEncapsulation } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { Component, effect, inject, ViewEncapsulation } from '@angular/core';
 import { MatButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { BaseDialogComponent } from 'src/app/core/dialogs/components/base-dialog/base-dialog.component';
 import { DialogComponent } from 'src/app/core/dialogs/components/dialog.component';
-import { DialogButtonBuilder } from 'src/app/core/dialogs/models/builder/dialog-button.builder';
 import { DialogBuilder } from 'src/app/core/dialogs/models/builder/dialog.builder';
 import { DialogConfiguration } from 'src/app/core/dialogs/models/configuration/dialog-configuration.model';
 import { DialogResolution } from 'src/app/core/dialogs/models/dialog-resolution.enum';
 import { EmailInputComponent } from 'src/app/core/firebase/components/email-input/email-input.component';
+import { FirebaseAuthDialogService } from 'src/app/core/firebase/components/firebase-auth-dialog/firebase-auth-dialog.service';
 import { PasswordInputComponent } from 'src/app/core/firebase/components/password-input/password-input.component';
-import { AuthScreen, AuthScreenUtility } from 'src/app/core/firebase/models/auth-screen.enum';
+import { AuthScreen } from 'src/app/core/firebase/models/auth-screen.enum';
 import { FirebaseService } from 'src/app/core/firebase/services/firebase.service';
-import { formValidity } from 'src/app/core/functions/signal/form-validity.function';
 import { EmptyComponent } from 'src/app/features/empty/empty.component';
 
 @Component({
@@ -20,157 +18,56 @@ import { EmptyComponent } from 'src/app/features/empty/empty.component';
 	templateUrl: 'firebase-auth-dialog.component.html',
 	styleUrl: 'firebase-auth-dialog.component.scss',
 	encapsulation: ViewEncapsulation.None,
+	providers: [ FirebaseAuthDialogService ],
 	imports: [
 		BaseDialogComponent, EmptyComponent,
 		MatButton, MatIcon, EmailInputComponent, PasswordInputComponent,
 	],
 })
 export class FirebaseAuthDialogComponent extends DialogComponent {
-	public readonly firebaseService = inject(FirebaseService);
+	private readonly firebaseAuthDialogService = inject(FirebaseAuthDialogService);
+	private readonly firebaseService = inject(FirebaseService);
 
 	public readonly AuthScreen = AuthScreen;
-	public readonly authScreen = signal(AuthScreen.AUTH_SELECTION);
+	public readonly authScreen = this.firebaseAuthDialogService.authScreen;
 
-	public readonly showBackButton = computed(() => AuthScreenUtility.canGoBack(this.authScreen()));
-	public readonly showCancelButton = computed(() => !this.showBackButton());
-
-	public readonly showNextButton = computed(() => AuthScreenUtility.canGoNext(this.authScreen()));
-	public readonly showEmailButtons = computed(() => AuthScreenUtility.canSubmit(this.authScreen()));
-
-	public readonly authSuccess = this.firebaseService.authSuccess;
-	public readonly authFailure = this.firebaseService.authFailure;
-
-	private readonly headerTitle = computed(() => {
-		if (this.authSuccess()) {
-			return 'Signed In';
-		} else if (this.authFailure()) {
-			return 'Sign In Failed';
-		} else {
-			return 'Sign In';
-		}
-	});
-
-	public readonly emailControl = new FormControl('', {
-		nonNullable: true,
-		validators: [
-			control => Validators.required(control),
-			control => Validators.email(control),
-		],
-	});
-	public readonly passwordControl = new FormControl('', {
-		nonNullable: true,
-		validators: [
-			control => Validators.required(control),
-		],
-	});
-	public readonly confirmPasswordControl = new FormControl('', {
-		nonNullable: true,
-		validators: [
-			control => Validators.required(control),
-		],
-	});
-
-	public readonly emailForm = new FormGroup({
-		email: this.emailControl,
-		password: this.passwordControl,
-		confirmPassword: this.confirmPasswordControl,
-	});
-
-	private readonly validEmail = formValidity(this.emailControl);
-	private readonly validPassword = formValidity(this.emailControl);
+	public readonly emailControl = this.firebaseAuthDialogService.emailControl;
+	public readonly passwordControl = this.firebaseAuthDialogService.passwordControl;
+	public readonly confirmPasswordControl = this.firebaseAuthDialogService.confirmPasswordControl;
 
 	public constructor() {
 		super();
 
-		effect(() => this.authSuccess() && (() => {
+		effect(() => this.firebaseService.authSuccess() && (() => {
 			this.authScreen.set(AuthScreen.SUCCESS);
 			setTimeout(() => this.dialogRef.close(DialogResolution.SUCCESS), 2_000);
 		})());
 
-		effect(() => this.authFailure() && (() => {
+		effect(() => this.firebaseService.authFailure() && (() => {
 			this.authScreen.set(AuthScreen.FAILURE);
 		})());
 	}
 
 	public initializeDialogConfiguration(): DialogConfiguration {
-		const backButton = DialogButtonBuilder.new()
-			.withAction(() => this.goBack())
-			.withText('Back')
-			.withIcon('arrow_left_alt')
-			.withVisible(this.showBackButton)
-			.withAlignment('left')
-			.build();
-
-		const nextButton = DialogButtonBuilder.new()
-			.withAction(() => this.goNext())
-			.withText('Next')
-			.withIcon('arrow_right_alt')
-			.withIconAlignment('right')
-			.withVisible(this.showNextButton)
-			.withEnabled(this.validEmail)
-			.withAlignment('right')
-			.build();
-
-		const signUpButton = DialogButtonBuilder.new()
-			.withAction(() => {})
-			.withText('Sign Up')
-			.withVisible(this.showEmailButtons)
-			.withAttribute('flat')
-			.withAlignment('right')
-			.withColor('primary')
-			.build();
-
-		const signInButton = DialogButtonBuilder.new()
-			.withAction(() => {})
-			.withText('Sign In')
-			.withVisible(this.showEmailButtons)
-			.withAlignment('right')
-			.withColor('primary')
-			.build();
+		const { title, buttons, showCancelButton } = this.firebaseAuthDialogService;
 
 		return DialogBuilder.new()
-			.withHeaderTitle(this.headerTitle)
+			.withHeaderTitle(title)
 			.withSubmitVisible(false)
-			.withCancelVisible(this.showCancelButton)
-			.withExtraButton(backButton)
-			.withExtraButton(nextButton)
-			.withExtraButton(signUpButton)
-			.withExtraButton(signInButton)
+			.withCancelVisible(showCancelButton)
+			.withExtraButtons(buttons)
 			.build();
 	}
 
 	public async selectGoogleSignIn(): Promise<void> {
-		this.authScreen.set(AuthScreen.GOOGLE_AUTH);
-		await this.firebaseService.googleSignIn();
+		await this.firebaseAuthDialogService.selectGoogleSignIn();
 	}
 
 	public selectEmailSignIn(): void {
-		this.authScreen.set(AuthScreen.EMAIL_AUTH);
-	}
-
-	private goBack(): void {
-		const current = this.authScreen();
-		const previous = AuthScreenUtility.getBackTarget(current);
-
-		if (previous !== null) {
-			this.emailForm.markAsPristine();
-			this.emailForm.markAsUntouched();
-			this.authScreen.set(previous);
-		}
-	}
-
-	private goNext(): void {
-		const current = this.authScreen();
-		const next = AuthScreenUtility.getNextTarget(current);
-
-		if (next !== null) {
-			this.emailForm.markAsPristine();
-			this.emailForm.markAsUntouched();
-			this.authScreen.set(next);
-		}
+		this.firebaseAuthDialogService.selectEmailSignIn();
 	}
 
 	public forgotPassword(): void {
-		this.authScreen.set(AuthScreen.RECOVER_PASSWORD);
+		this.firebaseAuthDialogService.forgotPassword();
 	}
 }
