+++ src/app/core/firebase/components/password-input/password-input.component.html
@@ -1,8 +1,13 @@
 <mat-form-field>
 	<mat-label>Password</mat-label>
 
-	<input matInput type="text" autocomplete="off"
+	<input matInput [type]="inputType()" autocomplete="off"
 		   [formControl]="passwordControl()"/>
+
+	<button type="button" mat-icon-button matIconSuffix
+			(click)="toggleVisibility()">
+		<mat-icon>{{ visibility() }}</mat-icon>
+	</button>
 
 	@if (passwordControl().getError('required')) {
 		<mat-error>Please enter your password.</mat-error>
